pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "xenium_alignment"
version = "0.1.0"
description = "Automated alignment tool between xenium and post-xenium"
requires-python = ">=3.9"
authors = [
    { name = "Hermet Leo", email = "Leo.HERMET@lyon.unicancer.fr" }
]
dependencies = [
    "geopandas>=0.14.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "shapely>=2.0.0",
    "SimpleITK>=2.3.0",
    "python-snappy>=0.6.1",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pyarrow>=12.0.0", 
    "tifffile>=2023.7.10",
    "ome-types>=0.4.0",
    "scikit-image>=0.20.0",
    "imagecodecs>=2023.1.23",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["xenium_align*"]